python-backend/requirements.txt
# QuantFusion v2.0 — Python Backend Dependencies
# Install: pip install -r requirements.txt

# — Core Server —
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=13.0
gunicorn>=21.2.0

# — Data APIs —
fyers-apiv3>=3.0
twelvedata>=1.2.0
httpx>=0.27.0
python-dotenv>=1.0.0
requests>=2.32.0

# — Token Management —
pyotp>=2.9.0
apscheduler>=3.10.0
pytz>=2024.1

# — Quant / Scientific —
numpy>=1.26.0
pandas>=2.2.0
scipy>=1.12.0
arch>=6.3.0

# — Machine Learning —
# stable-baselines3>=2.3.0  # Uncomment for RL Agent (requires PyTorch)
# gymnasium>=0.29.0         # Uncomment for RL Agent
scikit-learn>=1.5.0

# — NLP / Sentiment —
vaderSentiment>=3.3.2

# — Message Queue —
pyzmq>=25.1.0

# — Firebase —
firebase-admin>=6.5.0

# — Utilities —
pyyaml>=6.0.0
